Move and rename Air Source interface and backtank implementation, move static functions related to the backtank item.
