+++ src/main/java/com/simibubi/create/content/equipment/armor/DivingHelmetItem.java
@@ -1,5 +1,6 @@
 package com.simibubi.create.content.equipment.armor;
 
+import com.simibubi.create.content.equipment.armor.backtank_utils.IAirSource;
 import com.simibubi.create.foundation.advancement.AllAdvancements;
 
 import net.minecraft.resources.ResourceLocation;
@@ -77,7 +78,7 @@ public class DivingHelmetItem extends BaseArmorItem {
 		if (entity instanceof Player && ((Player) entity).isCreative())
 			return;
 
-		List<BacktankUtil.BacktankWrapper> backtanks = BacktankUtil.getAllWithAir(entity);
+		List<IAirSource> backtanks = BacktankUtil.getAllWithAir(entity);
 		if (backtanks.isEmpty())
 			return;
 
@@ -85,7 +86,7 @@ public class DivingHelmetItem extends BaseArmorItem {
 			if (entity instanceof ServerPlayer sp)
 				AllAdvancements.DIVING_SUIT_LAVA.awardTo(sp);
 			if (backtanks.stream()
-				.noneMatch(backtank -> backtank.isFireResistant()))
+				.noneMatch(IAirSource::isFireResistant))
 				return;
 		}
 
@@ -95,13 +96,12 @@ public class DivingHelmetItem extends BaseArmorItem {
 		if (world.isClientSide)
 			entity.getPersistentData()
 				.putInt("VisualBacktankAir", Math.round(backtanks.stream()
-					.map(BacktankUtil.BacktankWrapper::getAir)
+					.map(IAirSource::getAir)
 					.reduce(0f, Float::sum)));
 
 		if (!second)
 			return;
 
-//		BacktankUtil.consumeAir(entity, backtanks.get(0), 1);
 		backtanks.get(0).consumeAir(entity, 1);
 
 		if (lavaDiving)
